Command and event handling separated
