+++ src/index.js
@@ -21,26 +21,11 @@ client.player = new Player(client, {
     }
 });
 
-//event handling
-const eventFiles = fs.readdirSync('./events').filter(file => file.endsWith('.js'));
-for (const file of eventFiles) {
-    const event = require(`../events/${file}`);
-    if (event.once) {
-        client.once(event.name, (...args) => event.execute(...args));
-    } else {
-        client.on(event.name, (...args) => event.execute(...args));
-    }
+//function handling
+const functions = fs.readdirSync('./functions').filter(file => file.endsWith('.js'));
+for (const file of functions) {
+    require(`../functions/${file}`)(client);
 }
 
-//command handling
-client.commands = new discord.Collection();
-const commandFolders = fs.readdirSync('./commands');
-for (folder of commandFolders) {
-    const commandFiles = fs.readdirSync(`./commands/${folder}`).filter(file => file.endsWith('.js'));
-    for (file of commandFiles) {
-        const command = require(`../commands/${folder}/${file}`);
-        client.commands.set(command.data.name, command);
-    }
-}
 
 client.login(process.env.TOKEN);
